requirements.txt
Django==3.1.2
future==0.18.2
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.0.0
progress==1.5
pyparsing==2.4.6
PyPDF2==1.26.0
python-bidi==0.4.2
pytoml==0.1.21
pytz==2020.1
reportlab==3.5.53
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
Unipath==1.1
xhtml2pdf==0.2.5
django-model-utils==4.0.0